fix choppy timer

+++ StenGine/include/Utility/Timer.h
@@ -9,9 +9,9 @@ namespace StenGine
 
 class Timer {
 private:
-	static float m_gameStartTime;
-	static float m_lastUpdateTime;
-	static float m_deltaTime;
+	static double m_gameStartTime;
+	static double m_lastUpdateTime;
+	static double m_deltaTime;
 	static LARGE_INTEGER m_frequency;
 	static LARGE_INTEGER m_counter;
 
@@ -19,9 +19,9 @@ public:
 	Timer();
 	~Timer();
 	static void Init();
-	static float GetTime();
-	static float GetTimeSinceGameStart();
-	static float GetDeltaTime() { return m_deltaTime; }
+	static double GetTime();
+	static double GetTimeSinceGameStart();
+	static double GetDeltaTime() { return m_deltaTime; }
 	static void Update();
 };
