+++ StenGine/src/Utility/Timer.cpp
@@ -3,9 +3,9 @@
 namespace StenGine
 {
 
-float Timer::m_gameStartTime = 0;
-float Timer::m_lastUpdateTime = 0;
-float Timer::m_deltaTime = 0;
+double Timer::m_gameStartTime = 0;
+double Timer::m_lastUpdateTime = 0;
+double Timer::m_deltaTime = 0;
 LARGE_INTEGER Timer::m_frequency = { 0 };
 LARGE_INTEGER Timer::m_counter = { 0 };
 
@@ -15,17 +15,17 @@ void Timer::Init() {
 	m_lastUpdateTime = GetTime();
 }
 
-float Timer::GetTime() {
+double Timer::GetTime() {
 	QueryPerformanceCounter(&m_counter);
-	return (float)(m_counter.QuadPart) / (float)m_frequency.QuadPart;
+	return (double)(m_counter.QuadPart) / (double)m_frequency.QuadPart;
 }
 
-float Timer::GetTimeSinceGameStart() {
+double Timer::GetTimeSinceGameStart() {
 	return GetTime() - m_gameStartTime;
 }
 
 void Timer::Update() {
-	float time = GetTime();
+	double time = GetTime();
 	m_deltaTime = time - m_lastUpdateTime;
 	m_lastUpdateTime = time;
 }
